+++ packages/perspective-jupyterlab/test/jupyter/widget.spec.js
@@ -215,6 +215,7 @@ describe_jupyter(
 
                 // Check default config
                 expect(config).toEqual({
+                    version: utils.VIEWER_API_VERSION,
                     aggregates: {},
                     columns: [
                         "ui8",
@@ -267,6 +268,7 @@ w.theme = "Pro Dark"`
 
                 // and check it
                 expect(config).toEqual({
+                    version: utils.VIEWER_API_VERSION,
                     aggregates: {},
                     columns: ["ui8"],
                     expressions: [],
@@ -300,6 +302,7 @@ w.theme = "Pro Dark"`
 
                 // Check default config
                 expect(config).toEqual({
+                    version: utils.VIEWER_API_VERSION,
                     aggregates: {},
                     columns: [
                         "ui8",
