+++ rust/perspective-viewer/test/js/regressions.spec.js
@@ -12,6 +12,7 @@
 
 import { test, expect } from "@playwright/test";
 import {
+    VIEWER_API_VERSION,
     compareContentsToSnapshot,
     shadow_type,
 } from "@finos/perspective-test";
@@ -99,6 +100,7 @@ test.describe("Regression tests", () => {
         });
 
         expect(config).toEqual({
+            version: VIEWER_API_VERSION,
             aggregates: {},
             columns: ["Sales"],
             expressions: {},
